% Commit #2 % Kameron Damaska % 11/28/2016
Summary
=======

Two new files have been added to the Algorithms directory: HeapSort.java
and InsertionSort.java.

InsertionSort
-------------

The purpose of this class is to take a given array of integers and
return the array sorted using insertion sort.

The only method in this class is sort, which takes a parameter of
type Integer[] and returns a type Integer[].

HeapSort
--------

The purpose of this class is to take a given array of integers and
return the array sorted using heap sort.

The HeapSort class has the following private methods:

- leftChild: Returns the index of the left child for a given index.
- buildHeap: Constructs a max heap for a given array.
- bubbleUp:  Moves an integer at a given index to it's correct location
             in the heap.
- swap:      Swaps two values in an array.

The only public method in this class is sort, which takes a parameter of
type Integer[] and returns a type Integer[].